Create 11. *SoftUni Course Planning
Help plan the next Programming Fundamentals course by keeping track of the lessons that will be included in the course and all the exercises for the lessons. Before the course starts, there are some changes to be made.

On the first input line, you will receive the initial schedule of lessons and exercises that will be part of the next course, separated by a comma and a space ", ". Until you receive the "course start" command, you will be given some commands to modify the course schedule.

The possible commands are:

· "Add:{lessonTitle}" - add the lesson to the end of the schedule if it does not exist.

· "Insert:{lessonTitle}:{index}" - insert the lesson to the given index, if it does not exist.

· "Remove:{lessonTitle}" - remove the lesson, if it exists.

· "Swap:{lessonTitle}:{lessonTitle}" - swap the position of the two lessons if they exist.

· "Exercise:{lessonTitle}" - add Exercise in the schedule right after the lesson index, if the lesson exists and there is no exercise already, in the following format "{lessonTitle}-Exercise". If the lesson doesn't exist, add the lesson at the end of the course schedule, followed by the Exercise.